Update the codebase to use ES6 and support Node v4.3 on Lambda (#13)
Fixes #6

This required a few changes to the code base. First of all, we need to update our lambda function to use Node v4.3. This can be done in the Lambda console on AWS, and is also done in the default.env file. Next, since Node 4.3 doesn't offer full ES6 support yet, we need to run our source code through babel to transform it to be ES5-friendly. The ES6 source now lives in the src/ directory, and the transformed js is output into a dist/ directory. The dist/ package is what ends up getting uploaded to lambda.

One last issue is that node-lambda doesn't work with ES6 either, so we need to make sure it runs against the files in the dist/ directory instead. You can use the test-lambda npm script to do this automatically for you by executing npm run test-lambda in the terminal. This will trigger running babel to transform the js and then running node-lambda against the newly transformed files in dist/.