Output skill builder json from top interaction models.
The new skill builder has a slightly different JSON model that combines both the intent schema and the sample utterances/samples. This new generated JSON will allow you to upload the single file to the new skill builder and get both in one file. I'm guessing this will be the default way of setting the interaction model in the future.