Add support for displaying cards in the Alexa app (#10)
This requires providing an API key for TMDB, which is where we get the artwork from. You can do so in the config in the new `artwork` section:

```
  "server": {...},
  "movies": {...},
  "shows": {...},
  "artwork": {
    "tmdbApiKey": "abcdefghijklmnopqrstuvwxyz123456"
  }
```

Once configured correctly, a card will be displayed for Movies and Shows in the Alexa app (or in the Echo Show if you have one) after checking if a movie or show is already on your list, or after adding a new movie or show.