Add new config option for choosing a quality when adding media in sonarr/radarr (#11)
You can now optionally supply a key called "quality" for each provider that matches the quality you want the media to download at. Note that the name of the quality has to match exactly what it's called in Sonarr or Radarr for it to work. Example:

```
  "server": {...}
  "movies": {
    "provider": "radarr",
    "server": {...},
    "quality": "HD-1080p"
  },
  "shows": {
    "provider": "sonarr",
    "server": {...},
    "quality": "HDTV-1080p"
  }
```

If you don't set a quality, it'll use whatever the first quality profile from the API is, usually "Any".

Closes #6